+++ Source/Effects/Shards.c
@@ -53,7 +53,7 @@ typedef struct
 	OGLTextureCoord			uvs[3];
 	MOMaterialObject		*material;
 	OGLColorRGBA			colorFilter;
-	u_long					glow;
+	uint32_t					glow;
 }ShardType;
 
 
@@ -232,7 +232,7 @@ MOVertexArrayData	*vaData;
 static void ExplodeVertexArray(MOVertexArrayData *data, MOMaterialObject *overrideTexture)
 {
 OGLPoint3D			centerPt = {0,0,0};
-u_long				ind[3];
+uint32_t				ind[3];
 OGLTextureCoord		*uvPtr;
 long				i,t;
 float				boomForce = gBoomForce;
